requirements.txt
pyyaml
Flask
gunicorn
pymongo
pymongo[srv]
mongoengine
